Refactor ObjectLocation, StringUtils to io.lakefs.utils
