+++ clients/hadoopfs/src/main/java/io/lakefs/FileSystemTracer.java
@@ -1,5 +1,8 @@
 package io.lakefs;
 
+import io.lakefs.utils.ObjectLocation;
+import io.lakefs.utils.StringUtils;
+
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.*;
 import org.apache.hadoop.fs.permission.FsPermission;
