+++ clients/hadoopfs/src/main/java/io/lakefs/LakeFSFileSystem.java
@@ -3,6 +3,9 @@ package io.lakefs;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import io.lakefs.clients.api.*;
 import io.lakefs.clients.api.model.*;
+import io.lakefs.utils.ObjectLocation;
+import io.lakefs.utils.StringUtils;
+
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.*;
 import org.apache.hadoop.fs.permission.FsPermission;
@@ -845,35 +848,6 @@ public class LakeFSFileSystem extends FileSystem {
         }
     }
 
-    /**
-     * Returns Location with repository, ref and path used by lakeFS based on filesystem path.
-     *
-     * @param path to extract information from.
-     * @return lakeFS Location with repository, ref and path
-     */
-    @Nonnull
-    public ObjectLocation pathToObjectLocation(Path path) {
-        if (!path.isAbsolute()) {
-            path = new Path(this.workingDirectory, path);
-        }
-
-        URI uri = path.toUri();
-        ObjectLocation loc = new ObjectLocation();
-        loc.setScheme(uri.getScheme());
-        loc.setRepository(uri.getHost());
-        // extract ref and rest of the path after removing the '/' prefix
-        String s = StringUtils.trimLeadingSlash(uri.getPath());
-        int i = s.indexOf(Constants.SEPARATOR);
-        if (i == -1) {
-            loc.setRef(s);
-            loc.setPath("");
-        } else {
-            loc.setRef(s.substring(0, i));
-            loc.setPath(s.substring(i + 1));
-        }
-        return loc;
-    }
-
     class ListingIterator implements RemoteIterator<LakeFSFileStatus> {
         private final ObjectLocation objectLocation;
         private final String delimiter;
@@ -961,6 +935,10 @@ public class LakeFSFileSystem extends FileSystem {
                     objectLocation,
                     objectStats);
         }
+    }
+
+    ObjectLocation pathToObjectLocation(Path p) {
+        return ObjectLocation.pathToObjectLocation(workingDirectory, p);
     }
 
     private static boolean isDirectory(ObjectStats stat) {
