+++ clients/hadoopfs/src/main/java/io/lakefs/utils/StringUtils.java
@@ -1,15 +1,16 @@
-package io.lakefs;
+package io.lakefs.utils;
+
+import io.lakefs.Constants;
 
 public final class StringUtils {
-
-    static String trimLeadingSlash(String s) {
+    public static String trimLeadingSlash(String s) {
         if (s.startsWith(Constants.SEPARATOR)) {
             return s.substring(1);
         }
         return s;
     }
 
-    static String addLeadingSlash(String s) {
+    public static String addLeadingSlash(String s) {
         return s.isEmpty() || s.endsWith(Constants.SEPARATOR) ? s : s + Constants.SEPARATOR;
     }
 }
